extractor/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyGithub>=1.59.0
gitpython>=3.1.40
pydantic>=2.0.0
